requirements.txt
PyXB==1.2.4
lxml==3.6.0
requests==2.10.0
factory-boy==2.6.0
